Progress in this commit is all in the AVIReader input video file server. AVIReader will now import and serve cinepak compressed AVI files. Once AVIReader is running, you can drag such a file onto the media browser of DV, then drag onto the timeline tracks. You can also import bitmaps the same way.
As always, most of my changes are highlighted with a comment // ABH

So at this point you can finally import video content into UltraDV and
use a lot of the GUI interface features. There are still plenty of
crashes and testing has not really started.

Still don't know how to save a final video.

 Andrew